A couple more obvious missing (int64_t) cast fixes.

+++ modules/access/bda/bda.c
@@ -15,9 +15,9 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston MA 02110-1301, USA.
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston MA 02110-1301, USA.
  *****************************************************************************/
 
 /*****************************************************************************
@@ -556,7 +556,7 @@ static int Control( access_t *p_access, int i_query, va_list args )
         break;
     case ACCESS_GET_PTS_DELAY:      /* 5 */
         pi_64 = (int64_t*)va_arg( args, int64_t * );
-        *pi_64 = var_GetInteger( p_access, "dvb-caching" ) * 1000;
+        *pi_64 = (int64_t)var_GetInteger( p_access, "dvb-caching" ) * 1000;
         break;
         /* */
     case ACCESS_GET_TITLE_INFO:     /* 6 */
